Renamed io/bolt3.py to io/_bolt3.py
This is because this is an internal class implementing the Bolt
3 protocol.
