+++ neo4j/io/__init__.py
@@ -47,7 +47,7 @@ from time import perf_counter
 from neo4j.addressing import Address, AddressList
 from neo4j.api import ServerInfo
 from neo4j.conf import Config, PoolConfig
-from neo4j.io.bolt3 import Outbox, BufferedSocket, Inbox, Response, InitResponse, CommitResponse
+from neo4j.io._bolt3 import Outbox, BufferedSocket, Inbox, Response, InitResponse, CommitResponse
 from neo4j.errors import BoltRoutingError, Neo4jAvailabilityError
 from neo4j.exceptions import ProtocolError, SecurityError, \
     ServiceUnavailable, AuthError, IncompleteCommitError, \
